Make API load config as well.
